Create Exercise 27
(Replace words) Suppose you have a lot of files in a directory that contain words Exercisei_j, where i and j are digits. Write a program that pads a 0
before i if i is a single digit and 0 before j if j is a single digit. For example, the word Exercise2_1 in a file will be replaced by Exercise02_01. In Java, when
you pass the symbol * from the command line, it refers to all files in the directory (see Supplement III.V). Use the following command to run your program:
java Exercise12_27 *