Create Exercise 29
(Rename files) Suppose you have several files in a directory named Exercisei_j, where i and j are digits. Write a program that pads a 0 before j if j is a single
digit. For example, a file named Exercise2_1 in a directory will be renamed to Exercise2_01. In Java, when you pass the symbol * from the command line,
it refers to all files in the directory (see Supplement III.V). Use the following command to run your program:
java Exercise12_29 *