Create Exercise 04
Display calendars) Rewrite the PrintCalendar class in Listing 6.12 to display
a calendar for a specified month using the Calendar and GregorianCalendar classes. Your program receives the month and year from the command line. For
example: java Exercise13_04 5 2016 This displays the calendar shown in Figure 13.9.
You can also run the program without the year. In this case, the year is the current year. If you run the program without specifying a month and a year, the month is
the current month.