Create Exercise 06
(The ComparableCircle class) Define a class named ComparableCircle that extends Circle and implements Comparable. Draw the UML diagram
and implement the compareTo method to compare the circles on the basis of area. Write a test class to find the larger of two instances of ComparableCircle
objects, and the larger between a circle and a rectangle.