Create Exercise 12
(Sum the areas of geometric objects) Write a method that sums the areas of all the geometric objects in an array. The method signature is
public static double sumArea(GeometricObject[] a) Write a test program that creates an array of four objects (two circles and two rectangles) and computes their total area using the sumArea method.