Create Exercise 16
(Create a rational-number calculator) Write a program similar to Listing 7.9, Calculator.java. Instead of using integers, use rationals, as shown in Figure 13.10.
You will need to use the split method in the String class, introduced in Section 10.10.3, Replacing and Splitting Strings, to retrieve the numerator string
and denominator string, and convert strings into integers using the Integer.parseInt method.
(a) The program takes a string argument that consists of operand1, operator, and operand2 from the command line and displays the expression and the result of the arithmetic operation.