Create Exercise 21
(Algebra: vertex form equations) The equation of a parabola can be expressed in either standard form (y = ax2 + bx + c) or vertex form (y = a(x - h)2 + k).
Write a program that prompts the user to enter a, b, and c as integers in standard form and displays h  and k  in the vertex form. Display h and k as rational numbers. Here are some sample runs:
Enter a, b, c: 1 3 1
h is –3/2 k is –5/4

Enter a, b, c: 2 3 4
h is –3/4 k is 23/8