Update Exercise 03
(Display three cards) Write a program that displays three cards randomly selected from a deck of 52, as shown in Figure 14.43c. The card image files are
named 1.png, 2.png, . . . , 52.png and stored in the image/card directory. All three cards are distinct and selected randomly. (Hint: You can select random
cards by storing the numbers 1–52 to an array list, perform a random shuffle introduced in Section 11.12, and use the first three numbers in the array list as
the file names for the image.)