Create Exercise_05
(Display the divisors) Write a program that prompts the user to enter a positive integer and displays all its divisors in decreasing order. For example, if the integer is 12 the divisors are displayed as 6, 4, 3, 2, 1. Use the StackOfIntegers class to store the factors (e.g., 1, 2, 3, 4, 6) and retrieve and display them in reverse order.